+++ TestModels/dafny-dependencies/StandardLibrary/Makefile
@@ -3,6 +3,8 @@
 
 CORES=2
 
+TRANSPILE_TESTS_IN_RUST=true
+
 include ../../SharedMakefile.mk
 
 MAX_RESOURCE_COUNT=500000000
