+++ TestModels/dafny-dependencies/StandardLibrary/runtimes/rust/Cargo.toml
@@ -8,3 +8,6 @@ edition = "2021"
 [dependencies]
 dafny_runtime = { path = "../../../dafny_runtime_rust"}
 aws-smithy-types = "1.2.0"
+
+[lib]
+path = "src/implementation_from_dafny.rs"
