--- TestModels/dafny-dependencies/StandardLibrary/runtimes/rust/tests/standard_library_tests.rs
@@ -1,7 +1,0 @@
-
-mod tests_from_dafny;
-
-#[test]
-fn dafny_tests() {
-    crate::tests_from_dafny::_module::_default::_Test__Main_()
-}
